+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* API: station/changed does not have limit parameter
 ### Fixed
 * DB: database name is now allowed to be different from "radio"
 * CHECK: bitrate detection of stream works now on multiple bitrate headers